Fix system dimension outputting in a clumsy way
The idea is that the system should have some extension along z, which
the lattice box dimensions do not give. This should be solved by the
binary before output but that takes some more refactoring of that
utility, which is planned for when it does more things. For now the box
size along z is passed to the output function, which is stupid but
works for now.