Implement a Poisson Disc distribution using Bridson's algorithm
This is implemented alongside Lattices as another way of constructing a
substrate. The methods are of equal importance and thus a Points
construct has been added as the combined object which contains a
collection of points onto which Residues can be broadcast. This struct
now owns the method for uniformly distributing coordinates.

Fix Poisson distribution implementation

Add `Points` struct as a base for distributions and lattices

This struct defines a trait `IntoPoints` that should be used for any
type that creates a point collection. Additionally, methods like
`uniform_distribution` are moved to this struct.

Add PoissonDistribution as an option for substrate creation

The minimum distance is taken directly from the input density. This
needs to be tested to ensure that it works as intended.

Add randomized substrate to binary, with fixed settings

This should be user configurable, which is the next step for the binary.

Document new functions and modules