Make `Component` objects own their `ResidueBase`
Instead of struggling with keeping track of lifetimes it makes more
sense to do it like this. The original idea when building the
`System` which turned into a `Component` was that it would containt all
different residues, but this is not necessary. It is most likely easier
to have each component contain one and only one `ResidueBase` type.
These are then merged when finally outputting the system.

In addition of making it easier to reason about lifetimes, this means
that `Component`s are not bound to the `DataBase` definitions once they
have been created. This is good. And since each component still just
keeps a single `ResidueBase` object and their positions, little extra
memory is used.