Switch UI framework to `dialoguer`
Some functionality is not implemented yet, because `Atom` and
`ResidueBase` have to impl `Describe` in order to easily modify the list
of `ResidueBase` objects in the `DataBase`.

This should be done as the program is restructured, moving the
`DataBase` into the library part. No reason to make things more generic
than necessary.