Rework underlying component framework
The entire framework of creating components has been changed. Each
object that can be constructed as a component now only implements the
trait `Component`, not `IntoComponent`. They are then used indirectly by
both the `DataBase` and `System` objects through the superenum
`ComponentEntry` which wraps them to make them share one
`Vec<ComponentEntry>` object in the database or
construction for output: all kept in one place means less weird
overhead.

In addition, a macro has been added to easily implement this superenum:
`create_entry_wrapper`. This links required traits from each `Component`
for access.

Finally, the `System` is now defined to by default iterate over `Atom`s
using `iter_atoms`.

Add basic volume structures `Cuboid`, `Cylinder` and `Sphere`

Methods have been implemented to make `Cylinder`s and `Sphere`s from a
`Cuboid`. Traits `Periodic` and `Translate` have been added as suitable
for them.

Do not create documentation for the binary to avoid a name collision
with the library

[WIP] A new structure for components has been begun, with volumes as a
proof of concept

The idea is to not explicitly construct `Component`s. Instead, every
construction object (like a `Cuboid`) implements a trait (currently
`NewComponent`) which includes methods which iterate over all `Atom`s in
a system for when it is required for output.

This is done through an `AtomIterator` which is generic: it only
requires a `ResidueBase`, a set of `Coord`s for each residue and an
origin `Coord` for the component. A macro `impl_iter_atoms` has been
added to easily impl `NewComponent` for components and properly call this.

To move on, the `DataBase` entries should not be duplicated but only
one: each possible constructor has a single struct that calls
constructor methods on itself. These are then derived Deserialize and
Serialize for to read and write the database from and to disk. The
structure of using additional enum containers and match for these should
remain, since dynamic dispatch for these objects is impossible
(Serialize from serde seems to be the blocker). However, it should be
possible to use macros to quickly create these structures.

[WIP] Add macro `create_entry_wrapper` to generate a wrapping enum for constructors

Through this enum we access a common interface based on traits.

[WIP] Reimplement the `Sheet` and `Cylinder` surface as new constructors

[WIP] Add `fill` method to `volume::Cylinder` and add constructor traits to it

[WIP] Clean, then move new traits and macros

Note: binary does not currently implement the UI as the new structures
are being implemented

Changes include:
* Remove `Component` struct and all impls in `system`.
* Remove `AvailableComponent` from `database`.
* volume::{AtomIterator, CurrentAtom} -> iterator
* volume::NewComponent -> system::Component
* volume::Component now has trait boundaries

[WIP] `iter_atoms` methods return a Boxed type

This is to work more generally with iterator adaptors and is required to
implement the method for a system that combines them by chaining. This
has been added as a `System` struct with this method.

[WIP] Add `num_atoms` and `box_size` method to `Component` trait

[WIP] Implement `box_size` for `System` and add origin for it

[WIP] Rename `ResidueBase` to `Residue`

[WIP] Remove `System` from the binary code

This prepares for using the new `system::System`.

[WIP] Start reimplementing user interface

[WIP] Construct components now works

[WIP] Editing database reimplemented but untested

[WIP] Interface tested and working

[WIP] Update `database.json` to new format

[WIP] Extend `Describe` with `describe_short`

This is to have more control over the type of printed description of
objects.

[WIP] FIx warnings, add some #[allow(dead_code)] for future

[WIP] Get all user input through `get_value_from_user`

[WIP] Remove unused imports