Remove the `Rotate` trait
The reason is, no consistent design currently exists for having such a trait in the program. A rotation implies that an entire object is rotated: Right now, it only demanded that component (residue) components were rotated. Thus, a component could be rotated but its residues would still not change their orientation which is inconsistent.

It's also not easy to just rotate a component's residue if this is needed. If some transform has been applied to each residue in a component that might expect them to rotate individually, this is not covered by just rotating the shared residue coordinates. Instead, an individual rotation should be saved for each coordinate.

This can be done, and probably sometime the underlying coordinate lists should be able to have more explicit control over their residues, but that has not been designed yet. As such, it is currently dangerous to implement a `Rotate` trait that is dangerous and misguiding to use.